requirements.txt
defusedxml==0.4.1
Django==1.9.2
django-polymorphic==0.9.1
django-vanilla-views==1.0.4
djangorestframework==3.3.3
djangorestframework-xml==1.3.0
